requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
pymongo==4.7.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
